Move entities and dao
