--- app/shared/app-data/src/androidMain/kotlin/data/persistent/DatabaseStorage.android.kt
@@ -1,13 +1,0 @@
-package me.him188.ani.app.data.persistent
-
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import me.him188.ani.app.data.persistent.database.AniDatabase
-import me.him188.ani.app.platform.Context
-
-actual fun Context.createDatabaseBuilder(): RoomDatabase.Builder<AniDatabase> {
-    return Room.databaseBuilder<AniDatabase>(
-        context = applicationContext,
-        name = applicationContext.getDatabasePath("ani_room_database.db").absolutePath,
-    )
-}
