+++ app/shared/app-data/src/commonMain/kotlin/data/persistent/database/dao/EpisodeCollectionDao.kt
@@ -7,7 +7,7 @@
  * https://github.com/open-ani/ani/blob/main/LICENSE
  */
 
-package me.him188.ani.app.data.persistent.database
+package me.him188.ani.app.data.persistent.database.dao
 
 import androidx.paging.PagingSource
 import androidx.room.Dao
