+++ app/shared/app-data/src/commonMain/kotlin/data/persistent/database/dao/SearchHistoryDao.kt
@@ -1,10 +1,20 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.data.persistent.database.dao
 
 import androidx.room.Dao
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import androidx.room.Query
 import androidx.room.Upsert
 import kotlinx.coroutines.flow.Flow
-import me.him188.ani.app.data.persistent.database.eneity.SearchHistoryEntity
 
 @Dao
 interface SearchHistoryDao {
@@ -17,3 +27,9 @@ interface SearchHistoryDao {
     @Query("select * from `search_history` order by sequence desc")
     fun getFlow(): Flow<List<SearchHistoryEntity>>
 }
+
+@Entity(tableName = "search_history")
+data class SearchHistoryEntity(
+    @PrimaryKey(autoGenerate = true) val sequence: Int = 0,
+    val content: String
+)
