+++ app/shared/app-data/src/commonMain/kotlin/data/persistent/database/dao/SearchTagDao.kt
@@ -1,10 +1,20 @@
+/*
+ * Copyright (C) 2024 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
 package me.him188.ani.app.data.persistent.database.dao
 
 import androidx.room.Dao
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import androidx.room.Query
 import androidx.room.Upsert
 import kotlinx.coroutines.flow.Flow
-import me.him188.ani.app.data.persistent.database.eneity.SearchTagEntity
 
 @Dao
 interface SearchTagDao {
@@ -26,3 +36,13 @@ interface SearchTagDao {
     @Query("delete from `search_tag` where `id`=:id")
     suspend fun deleteById(id: Int)
 }
+
+@Entity(tableName = "search_tag")
+data class SearchTagEntity(
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val content: String,
+    /**
+     * 使用此 tag 搜索的次数，次数越高在搜索建议中排名越靠前
+     */
+    val useCount: Int = 1,
+)
