--- app/shared/app-data/src/commonMain/kotlin/data/persistent/database/eneity/SearchHistoryEntity.kt
@@ -1,10 +1,0 @@
-package me.him188.ani.app.data.persistent.database.eneity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "search_history")
-data class SearchHistoryEntity(
-    @PrimaryKey(autoGenerate = true) val sequence: Int = 0,
-    val content: String
-)
